refactor of AssignmentTexterContact splitting out the component from the container
