+++ src/lib/interaction-step-helpers.js
@@ -130,3 +130,37 @@ export function assembleAnswerOptions(allInteractionSteps) {
   });
   return interactionStepsCopy;
 }
+
+export function getAvailableInteractionSteps(
+  questionResponses,
+  allInteractionSteps
+) {
+  // helper for the client
+  // questionResponses: key=interactionStepId:value=questionResponse.value
+  // interactionSteps: all interaction steps
+  const availableSteps = [];
+
+  let step = getTopMostParent(allInteractionSteps);
+
+  while (step) {
+    availableSteps.push(step);
+    const questionResponseValue = questionResponses[step.id];
+    if (questionResponseValue) {
+      const matchingAnswerOption = step.question.answerOptions.find(
+        answerOption => answerOption.value === questionResponseValue
+      );
+      if (matchingAnswerOption && matchingAnswerOption.nextInteractionStep) {
+        step = interactionStepForId(
+          matchingAnswerOption.nextInteractionStep.id,
+          allInteractionSteps
+        );
+      } else {
+        step = null;
+      }
+    } else {
+      step = null;
+    }
+  }
+
+  return availableSteps;
+}
